requirements.txt
numpy
matplotlib
-e git+http://github.com/QUVA-Lab/artemis.git#egg=artemis
pytorch
